Adding common properties in "tapped_start" Firebase Event!
