+++ src/Firebase/base-firebase-integration.ts
@@ -1,5 +1,5 @@
 import { initializeApp } from "firebase/app";
-import { getAnalytics, logEvent ,setUserProperties} from "firebase/analytics";
+import { getAnalytics, logEvent, setUserProperties } from "firebase/analytics";
 import { firebaseConfig } from "./firebase-config";
 import { source, campaign_id } from "@common";
 
@@ -8,9 +8,9 @@ export class BaseFirebaseIntegration {
     analytics: any;
     constructor() {
         this.initializeFirebase();
-        console.log(" source : ",source ," and  campaign_id: ",campaign_id);
-        if(source!=null || campaign_id!=null)
-        this.setUserProperty(source ,campaign_id);
+        console.log(" source : ", source, " and  campaign_id: ", campaign_id);
+        if (source != null || campaign_id != null)
+            this.setUserProperty(source, campaign_id);
     }
     protected customEvents(eventName: string, event: object): void {
         try {
@@ -33,7 +33,7 @@ export class BaseFirebaseIntegration {
             setUserProperties(this.analytics, {
                 source: source,
                 campaign_id: campaignId
-            }, { global:true });
+            }, { global: true });
             console.log("User properties set: ", { source, campaignId });
         } catch (error) {
             console.error("Error while setting user properties:", error);
